Trinquete de salud y revisión de Copilot
El health gate falló por `archivos_800`: la ficha de rendimiento periódico
pasó de 622 a 858 líneas. Saco la gráfica SVG a
`ChartInteresesAcumulados.tsx` (657 + 217) — el indicador vuelve a 38 y la
ficha queda con su lógica, no con 180 líneas de path.

(`enlaces_rotos` 0→1 no es de este PR: es `/personal/resumen` en
`SubTabs.tsx`, ya exento por ampliación de definición.)

Revisión de Copilot, las cuatro válidas:
- El cuadro usaba `Math.round(duración / paso)`, que deja meses sin cubrir
  cuando la duración no es múltiplo de la frecuencia (10 meses en
  trimestral → 3 cuotas = 9 meses, con capital sin amortizar). Ahora
  `ceil`: la última cuota absorbe el periodo incompleto.
- `toDateInput` prometía cadena vacía ante fecha inválida y solo hacía
  `slice(0, 10)`, así que 'foo' se colaba en comparaciones de fechas.
  Valida el formato.
- `deletePosicion` es baja lógica, pero el toast decía "Posición
  eliminada" — justo lo contrario de lo que dice la zona peligrosa del
  wizard. Copy y mensaje de error alineados con el archivado.
